+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Holidays for New Zealand ([#1157])
+- Added grid into monthly calendar widget
 
 ### Changed
 - Updated holiday data